TP2

+++ SPE/C/TP2/pw_02_basics/power_of_two/power_of_two.c
@@ -1,3 +1,3 @@
 unsigned long power_of_two(unsigned char n) {
-    return 1 << n;
-}
+    return (unsigned long long)(1) << n;
+}
